% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep.R
\name{add_brackets}
\alias{add_brackets}
\title{Add Brackets}
\usage{
add_brackets(vector)
}
\arguments{
\item{vector}{A vector of values to add brackets around}
}
\value{
A character vector with brackets added
}
\description{
Adds brackets to input
}
\examples{
add_brackets(45)

}
