% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep.R
\name{add_top_row}
\alias{add_top_row}
\title{Add Top Row}
\usage{
add_top_row(data, value = NA_real_)
}
\arguments{
\item{data}{A data frame}

\item{value}{Value to fill the new row with (default: NA_real_)}
}
\value{
A data frame with an additional row at the top
}
\description{
Add an empty row on top of a \code{dataframe}
}
