% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep.R
\name{print_table}
\alias{print_table}
\title{Print Table}
\usage{
print_table(latex_table)
}
\arguments{
\item{latex_table}{A character vector representing a LaTeX table}
}
\value{
NULL (invisibly); prints to console
}
\description{
A cleaner way to print LaTeX tables than simply using the \code{cat()}.
Equivalent to \code{cat(paste(latex_table, collapse = "\\n"))}
}
\examples{
\dontrun{
table <- capture.output(kable(mtcars, format = "latex"))
print_table(table)
}
}
