% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep.R
\name{table_label}
\alias{table_label}
\title{Table Label}
\usage{
table_label(latex_table, caption)
}
\arguments{
\item{latex_table}{Table from \code{capture.output()}}

\item{caption}{Caption text for the table label}
}
\value{
A character vector representing the LaTeX table with label added
}
\description{
Adds labels to LaTeX tables that have are stored as a vector (e.g. output from \code{kable()} or \code{stargazer()} which is stored by \code{capture.output()}). Labels cannot be seen within the table,
but rather can be referenced within a rmarkdown document via \verb{\\label\{tab:your_label_here\}}
}
\examples{
\dontrun{
my_table_rough <- capture.output(stargazer(mtcars))
my_table <- table_label(my_table_rough, "demographics")
print_table(table)

# Within Markdown Document, reference via:
# `\label{tab:demographics}`
# This prints as "Table 1" and updates with table order.
}
}
