% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lm.R
\name{tidy_lm}
\alias{tidy_lm}
\title{tidy_lm (deprecated)}
\usage{
tidy_lm(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{tidy_lm()} was renamed to \code{tidy_lm_results()} to better reflect that it returns
a tidy data frame of results from multiple linear model specifications.
}
\keyword{internal}
